Refactor main.rb
